Rename package name in go files
