+++ pkg/vulcanizer/util.go
@@ -1,4 +1,4 @@
-package vulcan
+package vulcanizer
 
 import (
 	"strings"
